Modified to Copy Only Updates

+++ sync.sh
@@ -1,9 +1,8 @@
-
 #!/bin/bash
 
 
 #Enter Ur BackUp Folder Location
-BackUp_Dir="/home/paraqum/Paraqum/BackUp/"
+BackUp_Dir="/home/paraqum/Paraqum/BackUp_2/"
 
 # Store Tracked files by Git to y
 x="git ls-files"
@@ -12,21 +11,23 @@ y=$(eval "$x")
 # Get pwd and Display date
 Current_Dir=$(eval pwd)
 echo $(date)
-echo -e "\e[1;36m..Files and Subdirectories..\e[0m"
 
 # Acess the All Directories in the pwd
 cd $Current_Dir &&
 find . -not -name ".*" -type d -exec mkdir -p -- $BackUp_Dir{} \;
 
+echo -e "\e[1;36m..Synced Files and subdirectories..\e[0m"
+# Check whether same file exit in the BackUp Directory
+for item in $y; do
 
-for item in $y
-do
-	echo " $item"
-	cp -ar $item $BackUp_Dir/$item 
-		
+	if [[ -f $item ]] && [[ -f $BackUp_Dir/$item ]]; then
+    
+		# echo "$BackUp_Dir/$item"
+		cmp --silent -- $item $BackUp_Dir/$item || echo "$item is modified." && cp -ar $item $BackUp_Dir/$item
+
+    else
+	  	echo "$item is newly added to BackUp"
+		cp -ar $item $BackUp_Dir/$item 
+    fi
 done
-
-
-# Remove Hidden Directories
-#rm -r $BackUp_Dir.*
